+++ lib/generators/rodauth/views_generator.rb
@@ -55,6 +55,7 @@ module Rodauth
             copy_file "app/views/rodauth/#{view}.html.erb", "app/views/#{directory}/#{view}.html.erb" do |content|
               content = content.gsub("rodauth.", "rodauth(:#{configuration_name}).") if configuration_name
               content = content.gsub("rodauth/", "#{directory}/")
+              content = form_helpers_compatibility(content) if ActionView.version < Gem::Version.new("5.1")
               content
             end
           end
@@ -90,6 +91,16 @@ module Rodauth
 
         def configuration_name
           options[:name]&.to_sym
+        end
+
+        # We need to use the *_tag helpers on versions lower than Rails 5.1.
+        def form_helpers_compatibility(content)
+          content
+            .gsub(/form_with url: (.+) do \|form\|/, 'form_tag \1 do')
+            .gsub(/form\.(label|submit)/, '\1_tag')
+            .gsub(/form\.(email|password|text|telephone|hidden)_field (\S+), value:/, '\1_field_tag \2,')
+            .gsub(/form\.radio_button/, 'radio_button_tag')
+            .gsub(/form\.check_box (\S+), (.+) /, 'check_box_tag \1, "t", false, \2 ')
         end
       end
     end